feat: have in mind isRealDevice in utils.openFile
refactor to avoid some circular dependencies
