+++ nativescript-core/utils/native-helper.ios.ts
@@ -3,7 +3,8 @@ import {
     categories as traceCategories,
     write as traceWrite
 } from "../trace";
-import { isRealDevice } from "./utils";
+
+declare var UIImagePickerControllerSourceType: any;
 
 const radToDeg = Math.PI / 180;
 
@@ -15,7 +16,7 @@ function isOrientationLandscape(orientation: number) {
 function openFileAtRootModule(filePath: string): boolean {
     try {
         const appPath = ios.getCurrentAppPath();
-        let path = isRealDevice() ? filePath.replace("~", appPath) : filePath;
+        let path = ios.isRealDevice() ? filePath.replace("~", appPath) : filePath;
 
         const controller = UIDocumentInteractionController.interactionControllerWithURL(NSURL.fileURLWithPath(path));
         controller.delegate = new ios.UIDocumentInteractionControllerDelegateImpl();
@@ -152,6 +153,18 @@ export module ios {
 
         public documentInteractionControllerRectForPreview(controller: UIDocumentInteractionController): CGRect {
             return this.getViewController().view.frame;
+        }
+    }
+
+    export function isRealDevice() {
+        try {
+            // https://stackoverflow.com/a/5093092/4936697
+            const sourceType = UIImagePickerControllerSourceType.UIImagePickerControllerSourceTypeCamera;
+            const mediaTypes = UIImagePickerController.availableMediaTypesForSourceType(sourceType);
+
+            return mediaTypes;
+        } catch (e) {
+            return true;
         }
     }
 }
