+++ nativescript-core/utils/utils.android.ts
@@ -183,7 +183,6 @@ Please ensure you have your manifest correctly configured with the FileProvider.
     }
 }
 
-export function isRealDevice() {
-    const fingerprint = android.os.Build.FINGERPRINT;
-    this._isSimulator = fingerprint != null && (fingerprint.indexOf("vbox") > -1 || fingerprint.indexOf("generic") > -1);
+export function isRealDevice(): boolean {
+    return ad.isRealDevice();
 }
